pyproject.toml
[project]
name = "adk-customer-service"
version = "0.1.0"
description = "Customer Service (ADK)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.18.0",
]
